fix: refactor out getMatchingOption
