+++ src/components/fields/MultiSchemaField.js
@@ -6,8 +6,8 @@ import {
   getWidget,
   guessType,
   retrieveSchema,
+  getMatchingOption,
 } from "../../utils";
-import { isValid } from "../../validate";
 
 class AnyOfField extends Component {
   constructor(props) {
@@ -35,65 +35,11 @@ class AnyOfField extends Component {
 
   getMatchingOption(formData, options) {
     const { definitions } = this.props.registry;
-    for (let i = 0; i < options.length; i++) {
-      // Assign the definitions to the option, otherwise the match can fail if
-      // the new option uses a $ref
-      const option = Object.assign(
-        {
-          definitions,
-        },
-        options[i]
-      );
 
-      // If the schema describes an object then we need to add slightly more
-      // strict matching to the schema, because unless the schema uses the
-      // "requires" keyword, an object will match the schema as long as it
-      // doesn't have matching keys with a conflicting type. To do this we use an
-      // "anyOf" with an array of requires. This augmentation expresses that the
-      // schema should match if any of the keys in the schema are present on the
-      // object and pass validation.
-      if (option.properties) {
-        // Create an "anyOf" schema that requires at least one of the keys in the
-        // "properties" object
-        const requiresAnyOf = {
-          anyOf: Object.keys(option.properties).map(key => ({
-            required: [key],
-          })),
-        };
-
-        let augmentedSchema;
-
-        // If the "anyOf" keyword already exists, wrap the augmentation in an "allOf"
-        if (option.anyOf) {
-          // Create a shallow clone of the option
-          const { ...shallowClone } = option;
-
-          if (!shallowClone.allOf) {
-            shallowClone.allOf = [];
-          } else {
-            // If "allOf" already exists, shallow clone the array
-            shallowClone.allOf = shallowClone.allOf.slice();
-          }
-
-          shallowClone.allOf.push(requiresAnyOf);
-
-          augmentedSchema = shallowClone;
-        } else {
-          augmentedSchema = Object.assign({}, option, requiresAnyOf);
-        }
-
-        // Remove the "required" field as it's likely that not all fields have
-        // been filled in yet, which will mean that the schema is not valid
-        delete augmentedSchema.required;
-
-        if (isValid(augmentedSchema, formData)) {
-          return i;
-        }
-      } else if (isValid(options[i], formData)) {
-        return i;
-      }
+    let option = getMatchingOption(formData, options, definitions);
+    if (option !== 0) {
+      return option;
     }
-
     // If the form data matches none of the options, use the currently selected
     // option, assuming it's available; otherwise use the first option
     return this && this.state ? this.state.selectedOption : 0;
